+++ tests/org/elixir_lang/parser_definition/GeneralParsingTestCase.java
@@ -40,6 +40,14 @@ public class GeneralParsingTestCase extends ParsingTestCase {
         doTest(true);
     }
 
+    public void testSigilTerminatorEscape() {
+        doTest(true);
+    }
+
+    public void testValidEscapeSequences() {
+        doTest(true);
+    }
+
     @Override
     protected String getTestDataPath() {
         return super.getTestDataPath() + "/general_parsing_test_case";